updated bull functionality
