+++ services/workflows-service/src/bull-mq/webhook/types/types.ts
@@ -1,0 +1,9 @@
+import { Method } from 'axios';
+
+export interface WebhookJobData {
+  url: string;
+  method: Method;
+  headers?: Record<string, string>;
+  body?: Record<string, unknown>;
+  timeout?: number;
+}
